+++ RandomSkunk.Results/AsyncAction.cs
@@ -1,0 +1,7 @@
+namespace RandomSkunk.Results;
+
+/// <summary>
+/// Encapsulates an asynchronous method that has no parameters and does not return a value.
+/// </summary>
+/// <returns>A task that represents the asynchronous operation.</returns>
+public delegate Task AsyncAction();