+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning].
 - RandomSkunk.Results:
     - Add `[DebuggerDisplay]` to each of the result structs and the error class.
     - Add `CrossMap` extension methods to convert from one type of result to another.
+    - Add `ToResult`, `ToResult<T>`, `ToMaybe<T>`, `ToResultAsync`, `ToResultAsync<T>`, and `ToMaybeAsync<T>` extension methods to convert delegates into results.
 
 ### Fixed